SFT-2064: used composition instead of inheritance
