+++ ports/stm32/boards/Passport/manifest.py
@@ -151,7 +151,6 @@ freeze('$(MPY_DIR)/ports/stm32/boards/Passport/modules',
         'flows/sign_text_file_flow.py',
         'flows/terms_of_use_flow.py',
         'flows/update_firmware_flow.py',
-        'flows/uses_microsd_flow.py',
         'flows/verify_address_flow.py',
         'flows/verify_backup_flow.py',
         'flows/view_backup_code_flow.py',
